Refactor IORunnable wrapper
