+++ fahrschein/src/main/java/org/zalando/fahrschein/IORunnableAdapter.java
@@ -1,0 +1,20 @@
+package org.zalando.fahrschein;
+
+import java.io.IOException;
+
+final class IORunnableAdapter implements Runnable {
+    private final IORunnable delegate;
+
+    public IORunnableAdapter(IORunnable delegate) {
+        this.delegate = delegate;
+    }
+
+    @Override
+    public void run() {
+        try {
+            delegate.run();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
